+++ libs/nest_gui/gui_constraints.h
@@ -1,0 +1,19 @@
+#ifndef GUI_CONSTRAINTS
+#define GUI_CONSTRAINTS
+
+#include "nest.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif // !__cplusplus
+
+typedef struct _GUI_Constraint {
+    Nst_OBJ_HEAD;
+    Nst_GGC_HEAD;
+} GUI_Constraint;
+
+#ifdef __cplusplus
+}
+#endif // !__cplusplus
+
+#endif // !GUI_CONSTRAINTS