+++ libs/nest_gui/gui_draw.cpp
@@ -1,7 +1,7 @@
 #include "gui_draw.h"
 #include "gui_utils.h"
 
-void GUI_draw_rect(SDL_Renderer *renderer, SDL_Rect *rect)
+void GUI_DrawRect(SDL_Renderer *renderer, SDL_Rect *rect)
 {
     if (rect == nullptr)
         return;
@@ -77,9 +77,9 @@ static void draw_angle_border_point(SDL_Surface *surf, int rad, int border,
     }
 }
 
-SDL_Texture *GUI_draw_round_rect(SDL_Renderer *renderer, SDL_Rect rect,
-                                 int rtl, int rtr, int rbl, int rbr,
-                                 u8 r, u8 g, u8 b, u8 a)
+SDL_Texture *GUI_DrawRoundRect(SDL_Renderer *renderer, SDL_Rect rect,
+                               int rtl, int rtr, int rbl, int rbr,
+                               u8 r, u8 g, u8 b, u8 a)
 {
     SDL_Surface *surf = SDL_CreateRGBSurface(
         0, rect.w, rect.h, 32,
@@ -120,7 +120,7 @@ SDL_Texture *GUI_draw_round_rect(SDL_Renderer *renderer, SDL_Rect rect,
     cr = { rect.w - rbr, rect.h - rbr, rbr, rbr };
     SDL_FillRect(surf, &cr, 0);
 
-    int max_rad = max_int(max_int(rtl, rtr), max_int(rbl, rbr));
+    int max_rad = GUI_MaxInt2(GUI_MaxInt2(rtl, rtr), GUI_MaxInt2(rbl, rbr));
 
     for (int i = 0; i < max_rad; i++) {
         for (int j = 0; j < max_rad; j++) {
@@ -148,14 +148,14 @@ SDL_Texture *GUI_draw_round_rect(SDL_Renderer *renderer, SDL_Rect rect,
     return texture;
 }
 
-SDL_Texture *GUI_draw_round_border_rect(SDL_Renderer *renderer, SDL_Rect rect,
-                                        int border_thickness,
-                                        int rtl, int rtr, int rbl, int rbr,
-                                        u8 r_i, u8 g_i, u8 b_i, u8 a_i,
-                                        u8 r_b, u8 g_b, u8 b_b, u8 a_b)
+SDL_Texture *GUI_DrawRoundBorderRect(SDL_Renderer *renderer, SDL_Rect rect,
+                                     int border_thickness,
+                                     int rtl, int rtr, int rbl, int rbr,
+                                     u8 r_i, u8 g_i, u8 b_i, u8 a_i,
+                                     u8 r_b, u8 g_b, u8 b_b, u8 a_b)
 {
     if (border_thickness == 0) {
-        return GUI_draw_round_rect(
+        return GUI_DrawRoundRect(
             renderer,
             rect,
             rtl, rtr, rbl, rbr,
@@ -208,7 +208,7 @@ SDL_Texture *GUI_draw_round_border_rect(SDL_Renderer *renderer, SDL_Rect rect,
     cr = { rect.w - rbr, rect.h - rbr, rbr, rbr };
     SDL_FillRect(surf, &cr, 0);
 
-    int max_rad = max_int(max_int(rtl, rtr), max_int(rbl, rbr));
+    int max_rad = GUI_MaxInt2(GUI_MaxInt2(rtl, rtr), GUI_MaxInt2(rbl, rbr));
 
     for (int i = 0; i < max_rad; i++) {
         for (int j = 0; j < max_rad; j++) {
@@ -244,7 +244,7 @@ SDL_Texture *GUI_draw_round_border_rect(SDL_Renderer *renderer, SDL_Rect rect,
     return texture;
 }
 
-void GUI_draw_texture(SDL_Renderer *renderer, int x, int y,
+void GUI_DrawTexture(SDL_Renderer *renderer, int x, int y,
                       SDL_Texture *texture, SDL_Rect *clip)
 {
     int w, h;
@@ -259,6 +259,9 @@ void GUI_draw_texture(SDL_Renderer *renderer, int x, int y,
     SDL_Rect dst_rect;
     if (!SDL_IntersectRect(clip, &texture_rect, &dst_rect))
         return;
-    SDL_Rect src_rect = { dst_rect.x - x, dst_rect.y - y, dst_rect.w, dst_rect.h };
+    SDL_Rect src_rect = {
+        dst_rect.x - x, dst_rect.y - y,
+        dst_rect.w, dst_rect.h
+    };
     SDL_RenderCopy(renderer, texture, &src_rect, &dst_rect);
 }