+++ libs/nest_gui/gui_draw.h
@@ -8,16 +8,16 @@
 extern "C" {
 #endif // !__cplusplus
 
-void GUI_draw_rect(SDL_Renderer *renderer, SDL_Rect *rect);
-SDL_Texture *GUI_draw_round_rect(SDL_Renderer *renderer, SDL_Rect rect,
+void GUI_DrawRect(SDL_Renderer *renderer, SDL_Rect *rect);
+SDL_Texture *GUI_DrawRoundRect(SDL_Renderer *renderer, SDL_Rect rect,
                                  int rtl, int rtr, int rbl, int rbr,
                                  u8 r, u8 g, u8 b, u8 a);
-SDL_Texture *GUI_draw_round_border_rect(SDL_Renderer *renderer, SDL_Rect rect,
+SDL_Texture *GUI_DrawRoundBorderRect(SDL_Renderer *renderer, SDL_Rect rect,
                                         int border_thickness,
                                         int rtl, int rtr, int rbl, int rbr,
                                         u8 r_i, u8 g_i, u8 b_i, u8 a_i,
                                         u8 r_b, u8 g_b, u8 b_b, u8 a_b);
-void GUI_draw_texture(SDL_Renderer *renderer, int x, int y,
+void GUI_DrawTexture(SDL_Renderer *renderer, int x, int y,
                       SDL_Texture *texture, SDL_Rect *clip);
 
 #ifdef __cplusplus