+++ libs/nest_gui/nest_gui.h
@@ -2,6 +2,9 @@
 #define NEST_GUI_H
 
 #include "nest.h"
+#include "gui_app.h"
+#include "gui_draw.h"
+#include "gui_element.h"
 #include "gui_utils.h"
 
 #ifdef Nst_WIN
@@ -18,6 +21,15 @@ NstEXP void NstC free_lib();
 
 Nst_FUNC_SIGN(init_sdl_and_ttf_);
 Nst_FUNC_SIGN(loop_);
+Nst_FUNC_SIGN(window_init_);
+Nst_FUNC_SIGN(window_set_title_);
+Nst_FUNC_SIGN(window_get_title_);
+Nst_FUNC_SIGN(window_set_position_);
+Nst_FUNC_SIGN(window_get_position_);
+Nst_FUNC_SIGN(window_set_size_);
+Nst_FUNC_SIGN(window_get_size_);
+Nst_FUNC_SIGN(window_set_resizable_);
+Nst_FUNC_SIGN(window_get_resizable_);
 
 #ifdef __cplusplus
 }